[refector] (bitcoin) move Singlesig functions to bitcoinish
