+++ packages/coinlib-bitcoin/src/bitcoinish/index.ts
@@ -6,3 +6,4 @@ export * from './helpers'
 export * from './utils'
 export * from './BitcoinishBalanceMonitor'
 export * from './multisigPaymentHelper'
+export * from './singlesigPaymentHelper'
